+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 All notable changes to `unit-conversion` will be documented in this file.
 
 # 03/07/2023
-UnitConverters for Temperature groups added.
+- UnitConverters for Temperature groups added.
+- Weight Conversion class (Kilogram) added.